Move backend parameter definitions from server.pgcluster to pgsql
Backend capabilities and other parameters aren't exclusively a server
concern, so move them to a more generic spot.
